refactor cutpoint workflow. avoid bug with log axis. fix #96
